requirements.txt
pandas==2.2.2
numpy==1.26.4
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
scipy==1.13.1
statsmodels==0.14.2
scikit-learn==1.5.0
xgboost==2.0.3
optuna==3.6.1
shap==0.45.1
mlflow==2.14.1
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
streamlit==1.35.0
plotly==5.22.0
matplotlib==3.9.0
joblib==1.4.2
pytest==8.2.2
